% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/environment.R
\name{as.mp_options}
\alias{as.mp_options}
\title{FUNCTION_TITLE}
\usage{
as.mp_options(x = NULL)
}
\arguments{
\item{x}{PARAM_DESCRIPTION, Default: NULL}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
