% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posterior.R
\name{calc_diagnostics}
\alias{calc_diagnostics}
\title{FUNCTION_TITLE}
\usage{
calc_diagnostics(iw, include_extras = FALSE)
}
\arguments{
\item{iw}{PARAM_DESCRIPTION}

\item{include_extras}{Whether to include transformed quantities used to
calculate the main quantities, Default: FALSE}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
