% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posterior.R
\name{calc_expectation}
\alias{calc_expectation}
\title{FUNCTION_TITLE}
\usage{
calc_expectation(imp_weights, quantity)
}
\arguments{
\item{imp_weights}{PARAM_DESCRIPTION}

\item{quantity}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
