% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posterior.R
\name{calc_imp_weights}
\alias{calc_imp_weights}
\title{FUNCTION_TITLE}
\usage{
calc_imp_weights(log_r, method = c("psis", "is", "average"))
}
\arguments{
\item{log_r}{PARAM_DESCRIPTION}

\item{method}{PARAM_DESCRIPTION, Default: "psis"}
}
\value{
\code{positive}, \code{log_r}, \code{log_w}, \code{r}, \code{w}
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[matrixStats]{logSumExp}}
}
