% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posterior.R
\name{calc_importance_sampling}
\alias{calc_importance_sampling}
\title{FUNCTION_TITLE}
\usage{
calc_importance_sampling(estimate, N_mc, options, ...)
}
\arguments{
\item{estimate}{PARAM_DESCRIPTION}

\item{N_mc}{PARAM_DESCRIPTION}

\item{options}{PARAM_DESCRIPTION}

\item{...}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
