% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform.R
\name{dwishart}
\alias{dwishart}
\title{FUNCTION_TITLE}
\usage{
dwishart(
  W = NULL,
  x = NULL,
  W_chol = NULL,
  V_chol,
  df,
  lower_chol = FALSE,
  log = FALSE
)
}
\arguments{
\item{W}{PARAM_DESCRIPTION, Default: NULL}

\item{x}{PARAM_DESCRIPTION, Default: NULL}

\item{W_chol}{PARAM_DESCRIPTION, Default: NULL}

\item{V_chol}{PARAM_DESCRIPTION}

\item{df}{PARAM_DESCRIPTION}

\item{lower_chol}{PARAM_DESCRIPTION, Default: FALSE}

\item{log}{PARAM_DESCRIPTION, Default: FALSE}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
This function requires the \code{CholWishart::lmvgamma} function from
the \code{CholWishart} package.
}
\examples{
\dontrun{
if (interactive()) {
  # EXAMPLE1
}
}
}
