% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{inverse_chol_reverse}
\alias{inverse_chol_reverse}
\title{Inverse Reverse Cholesky}
\usage{
inverse_chol_reverse(Sigma_chol, lower_chol = FALSE)
}
\arguments{
\item{Sigma_chol}{Holesky factor of a covariance matrix}

\item{lower_chol}{\code{TRUE} if lower triangular Cholesky factors are used
Default: FALSE}
}
\value{
A Cholesky matrix of the same triangle orientation as the input,
with rows and columns in reverse order.
}
\description{
Compute inverse of a Cholesky matrix,
storing the result in reverse order, see Details.
}
\details{
For \code{lower_chol == FALSE},
the input \code{Sigma_chol} is the matrix \eqn{R} in the Cholesky factorisation
\eqn{\Sigma=R^T R} with inverse \eqn{Q=R^{-1}R^{-T}}. Since \eqn{R^{-1}} has
the opposite upper/lower triangular property to \eqn{R^T}, this \eqn{Q}
factorisation isn't a regular Cholesky factorisation. Let \eqn{P} be the
permutation matrix that reverses element order. Then
\eqn{PQP=PR^{-1}PPR^{-T}P}, and \eqn{PR^{-T}P} is the Cholesky factor of
\eqn{PQP} with the same upper/lower triangular property as \eqn{R}.

For \code{lower_chol == TRUE}, the input is \eqn{L} in \eqn{\Sigma=LL^T} and the
output is \eqn{PL^{-T}P}, with \eqn{PQP=PL^{-T}PPL^{-1}P}.
}
\examples{
if(interactive()){
  inverse_chol_reverse(matrix(c(1, 0, 2, 3), 2, 2))
}
}
\keyword{internal}
