% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform.R
\name{latent_from_wishart}
\alias{latent_from_wishart}
\title{Transform Wishart to latent variables}
\usage{
latent_from_wishart(LC, LV, df)
}
\arguments{
\item{LC}{The lower Cholesky factor of the Wishart matrix}

\item{LV}{The lower Cholesky factor of the Wishart \eqn{V} parameter}

\item{df}{The Wishart degrees of freedom}
}
\value{
A list
}
\description{
Transforms a Wishart matrix to latent iid \eqn{N(0,1)} variables
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\author{
Finn Lindgren
}
