% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform.R
\name{latent_to_nwishart}
\alias{latent_to_nwishart}
\title{Transform latent variables to Normalised Wishart}
\usage{
latent_to_nwishart(x, LV, df)
}
\arguments{
\item{x}{A vector of latent variables, length $d(d-1)/2$}

\item{LV}{The lower Cholesky factor of the Wishart $V$ parameter}

\item{df}{The Wishart degrees of freedom}
}
\value{
A list
}
\description{
Transforms latent iid $N(0,1)$ to Normalised Wishart matrices
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\author{
Finn Lindgren
}
