% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform.R
\name{latent_to_wishart}
\alias{latent_to_wishart}
\title{Transform latent variables to Wishart}
\usage{
latent_to_wishart(x, LV, df)
}
\arguments{
\item{x}{A vector of latent variables, length \eqn{d(d+1)/2}}

\item{LV}{The lower Cholesky factor of the Wishart \eqn{V} parameter}

\item{df}{The Wishart degrees of freedom}
}
\value{
A list
}
\description{
Transforms latent iid \eqn{N(0,1)} to Wishart matrices
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\author{
Finn Lindgren
}
