% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posterior.R
\name{log_psis_raw}
\alias{log_psis_raw}
\title{Pareto Smoothed Importance Sampling Weights}
\usage{
log_psis_raw(log_r)
}
\arguments{
\item{log_r}{logarithm of importance ratios for importance sampling}
}
\value{
Pareto smoothed importance log-ratios. \code{-Inf} values are converted
to \verb{-log(.Machine$double.eps}, and finite values are fed through \code{loo::psis},
and the output is shifted to match the original offset. \code{loo::psis} on
its own shifts the log-ratios by the maximum input value, but we need the
offset to be the same as in the input, so that the resulting
\eqn{z^*=\sum_k r_k} doesn't change its interpretation before and after
applying the pareto smoothing.
}
\description{
Wrapper for \code{loo::psis} with \code{multiprobit} specific
requirements handled.
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[loo]{weights.importance_sampling}}
}
