% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/environment.R
\name{mp_log_message}
\alias{mp_log_message}
\title{FUNCTION_TITLE}
\usage{
mp_log_message(..., domain = NULL, appendLF = TRUE, allow_verbose = TRUE)
}
\arguments{
\item{...}{zero or more objects passed on to \code{\link[base:.makeMessage]{base::.makeMessage()}}}

\item{domain}{Domain for translations, passed on to \code{\link[base:.makeMessage]{base::.makeMessage()}}}

\item{appendLF}{logical; whether to add a newline to the message. Only
used for verbose output.}

\item{allow_verbose}{Whether to allow verbose output. Must be set to FALSE
until the options object has been initialised.}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if (interactive()) {
  # EXAMPLE1
}
}
}
