% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimation.R
\name{mp_model}
\alias{mp_model}
\title{Setup multivariate probit model}
\usage{
mp_model(
  model = NULL,
  response = NULL,
  X = NULL,
  formula = NULL,
  data = NULL,
  df = NULL,
  prec_beta = NULL
)
}
\arguments{
\item{model}{An optional existing model object to be updated.}

\item{response}{A matrix with n-by-d elements, where each row is a
multivariate observation, see Details. A vector is interpreted as a
single row matrix.}

\item{X}{An optimally precomputed n-by-J model matrix, where J is the number
regression coeficcients for each of the d dimensions.}

\item{formula}{A formula interpretable by \code{model.matrix}.}

\item{data}{A \code{data.frame} containing the veriables needed by the
formula.}

\item{df}{Degrees of freedom for the normalised Wishart prior for the
correlation matrix. See Details.}

\item{prec_beta}{Prior precision for the regression coefficients}
}
\value{
An object of class \code{mp_model}
}
\description{
Construct a multivariate probit model object that stores
information about model structure and parameters
}
\details{
The multivariate probit model has a multivariate binary
response variable, here denoted \eqn{Y}. The model is built from a
linear predictor
\deqn{M = X B}
where \eqn{X} is a n-by-J matrix of \eqn{J} predictors, and \eqn{B} is
a J-by-d matrix of regression coefficients.
Each row of \eqn{M} is the linear
predictor for one multivariate observation. The response variables \eqn{Y}
are linked to \eqn{M} by first defining latent Gaussian variables
\deqn{Z=M+E} where each row of \eqn{E} is a multivariate Normal vector,
\eqn{E \sim N(0,\Sigma)}. Then, \deqn{Y_{i,k}=I(Z_{i,k} > 0).}
Conditionally on \eqn{B}, each row of \eqn{Y} has a multinomial distribution
on the set of all \eqn{0/1} combinations, with each probability equal to a
hyperquadrant probability of a the multivariate Normal distribution
\eqn{N(\mu,\Sigma)}, where \eqn{\mu} is the corresponding row of \eqn{M}.

Only the inequality \eqn{Y_{i,k} > 0} for the response variables is used,
so alternative data representations such as \eqn{-1/+1} will also work as
expected.

The degrees of freedom for the normalised Wishart prior are linked to the
LKJ prior by \code{df = ...}, which makes the two models equivalent.
}
\examples{
\dontrun{
if (interactive()) {
  # EXAMPLE1
}
}
}
