% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/environment.R
\name{mp_options_check}
\alias{mp_options_check}
\title{Check multiprobit options}
\usage{
mp_options_check(options)
}
\arguments{
\item{options}{The \code{\link{mp_options}} object to check}
}
\value{
Logical; \code{TRUE} if the object is valid
}
\description{
Check for valid contents of an mp_options object
}
\details{
Warnings are produced for invalid contents
}
\examples{
\dontrun{
if(interactive()){
  mp_options_check(mp_options(invalid = "something"))
}
}
}
