% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/environment.R
\name{mp_options_default}
\alias{mp_options_default}
\title{FUNCTION_TITLE}
\usage{
mp_options_default()
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if (interactive()) {
  # EXAMPLE1
}
}
}
