% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/environment.R
\name{mp_options_get}
\alias{mp_options_get}
\title{FUNCTION_TITLE}
\usage{
mp_options_get(x = NULL)
}
\arguments{
\item{x}{PARAM_DESCRIPTION, Default: NULL}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if (interactive()) {
  # EXAMPLE1
}
}
}
