% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/environment.R
\name{mp_options_merge}
\alias{mp_options_merge}
\title{FUNCTION_TITLE}
\usage{
mp_options_merge(...)
}
\arguments{
\item{...}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
This mid-level function does not access the global and/or default options
by itself.
}
\examples{
\dontrun{
if (interactive()) {
  # EXAMPLE1
}
}
}
