% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/environment.R
\name{mp_options_set}
\alias{mp_options_set}
\title{Set global multiprobit options}
\usage{
mp_options_set(..., default = FALSE)
}
\arguments{
\item{...}{Zero or more global multiprobit options, see \code{\link[=mp_options]{mp_options()}} for details of
available options.}

\item{default}{logical; If \code{TRUE}, reinitialise the global options with
default options before setting new options. Default: \code{FALSE}}
}
\value{
A copy of the global options is returned invisibly.`
}
\description{
Changes the global options. In many cases, setting options
in specific calls to \code{\link[=multiprobit]{multiprobit()}} is recommended instead.
}
\examples{
\dontrun{
if (interactive()) {
  mp_options_set(gauusint = list(max.threads = 1),
                 optim = list(trace = 5))
}
}
}
\seealso{
\code{\link[=mp_options]{mp_options()}}, \code{\link[=mp_options_default]{mp_options_default()}}, \code{\link[=mp_options_merge]{mp_options_merge()}},
\code{\link[=mp_options_get]{mp_options_get()}}
}
