% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/probabilities.R
\name{mpp}
\alias{mpp}
\title{Multivariate Probit Event Probabilities}
\usage{
mpp(
  y,
  mu,
  Sigma_chol = NULL,
  Q_chol = NULL,
  log = FALSE,
  lower_chol = FALSE,
  seed = NULL,
  ...
)
}
\arguments{
\item{y}{A matrix with n-by-d elements, where each row is a multivariate
observation, see Details. A vector is interpreted as a single row matrix.}

\item{mu}{A matrix with n-by-d elements, where each row is an expectation
vector parameter, see Details. A vector is interpreted as a single row
matrix.}

\item{Sigma_chol}{The Cholesky factor of the covariance matrix parameter,
Default: NULL}

\item{Q_chol}{The Cholesky factor of the precision matrix matrix parameter,
Default: NULL}

\item{log}{logical indicating if the log-probability should be returned, Default: FALSE}

\item{lower_chol}{\code{TRUE} if lower triangular Cholesky factors are used
Default: FALSE}

\item{seed}{The random seed for \code{excursions::gaussint}, Default: NULL}

\item{...}{Further parameters, ignored.}
}
\value{
A list with components
\describe{
\item{P}{A vector of probabilities}
\item{E}{A vector with the estimated approximation error for each probability}
}
}
\description{
Evaluate one or more probabilities for outcomes of a
multivariate probit model, with given location and covariance scale
parameteters.
}
\details{
Computes the probability
  \deqn{P(y_1 > 0, ..., y_d > 0|\mu,\Sigma)}
when \eqn{y} is a \eqn{d}-dimensional indicator vector with elements
\eqn{y_i=I(z_i \ge 0)}, and \eqn{z} is a \eqn{d}-dimensional
Gaussian vector with distribution \eqn{N(\mu,\Sigma)}. Only the inequality
for \eqn{y_i} is used, so alternative data representations such as
\eqn{-1/+1} will also work as expected.

The \eqn{\Sigma} paramter can either be specified though its Cholesky factor
\code{Sigma_chol} or through the Cholesky factor of the precision (inverse
of \eqn{\Sigma}) \code{Q_chol}.
The logical parameter \code{lower_chol} determines if a lower or upper
triangular Cholesky factor was supplied.

The \code{seed} parameter is passed on to \code{excursions::gaussint}, and
allows consistent approximation error when calculating numerical derivatives.
}
\examples{
if (interactive()) {
  mpp(
    y = c(1, 0),
    mu = c(1, 2),
    Sigma_chol = chol(matrix(c(1, 0.5, 0.5, 1), 2, 2))
  )
}
}
\seealso{

}
