% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/probabilities.R
\name{mpp_gradient_mu}
\alias{mpp_gradient_mu}
\alias{mpp_hessian_mu}
\alias{mpp_gradient_u}
\title{Multivariate Probit Event Probability Derivatives}
\usage{
mpp_gradient_mu(
  y,
  mu,
  ...,
  gaussint_options = NULL,
  h = 1e-06,
  symmetric = FALSE
)

mpp_hessian_mu(
  y,
  mu,
  ...,
  gaussint_options = NULL,
  h = 1e-04,
  diagonal = FALSE
)

mpp_gradient_u(
  y,
  mu,
  u,
  Sigma_model,
  gaussint_options = NULL,
  h = 1e-06,
  symmetric = FALSE,
  log = FALSE,
  ...
)
}
\arguments{
\item{y}{A matrix of multivariate 0/1 observations}

\item{mu}{A matrix of matrix of multivariate latent scale expectation
parameters}

\item{...}{Further parameters passed on to
\code{\link[=mpp]{mpp()}}}

\item{gaussint_options}{list of options for \code{excursions::gaussint}.
By default sets \code{seed = 1L} to ensure consistent approximation error.}

\item{h}{Step size for finite differences, Default: 1e-06 (for gradients)
or 1e-04 (for hessian)}

\item{symmetric}{For gradients, whether to use symmetric finite differences,
Default: FALSE}

\item{diagonal}{Logical; if \code{TRUE}, only the diagonal of the hessian
is evaluated, Default: FALSE}

\item{u}{A vector of latent variables identifying the Normalised Wishart
matrix, length \eqn{d(d-1)/2}}

\item{Sigma_model}{A \code{\link{wm_model}} object}

\item{log}{Whether to compute gradient of the log-probability,
Default: \code{FALSE}}
}
\value{
OUTPUT_DESCRIPTION gradient for mu

OUTPUT_DESCRIPTION hessian for mu

OUTPUT_DESCRIPTION gradient for u
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS gradient

DETAILS hessain

DETAILS
}
\examples{
\dontrun{
if (interactive()) {
  # EXAMPLE1 gradient
}
}
\dontrun{
if (interactive()) {
  # EXAMPLE1 hessian
}
}
\dontrun{
if (interactive()) {
  # EXAMPLE1
}
}
}
\seealso{
\code{\link[=mpp]{mpp()}}
}
