% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiprobit-package.R
\docType{package}
\name{multiprobit-package}
\alias{multiprobit-package}
\alias{_PACKAGE}
\title{multiprobit: Bayesian Multivariate Probit Regression}
\description{
Perform fast Bayesian inference for multivariate
    probit models.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/finnlindgren/multiprobit}
  \item Report bugs at \url{https://github.com/finnlindgren/multiprobit/issues}
}

}
\author{
\strong{Maintainer}: Finn Lindgren \email{finn.lindgren@gmail.com}

}
\keyword{internal}
