% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimation.R
\name{multiprobit}
\alias{multiprobit}
\title{Estimate a multivariate probit model}
\usage{
multiprobit(
  response = NULL,
  model,
  gaussint_options = NULL,
  optim_options = NULL,
  max_iter = 10,
  method = c("alternating", "joint", "stepwise")
)
}
\arguments{
\item{response}{A matrix with n-by-d elements, where each row is a multivariate
observation, see Details. A vector is interpreted as a single row matrix.
If \code{NULL}, the observations need to be present in the \code{model}
object.}

\item{model}{An \code{\link{mp_model}} model object.}

\item{gaussint_options}{Optional list of options for
\code{excursions::gaussint}.
Specific relevant option:
\code{num_threads} The maximum number of allowed threads for parallel
computing by \code{excursions::gaussint}, Default: 0, meaning no limit.}

\item{optim_options}{Optional list of control options for \code{optim()}}

\item{max_iter}{The maximum number of steps for
\code{method == "alternating"}}

\item{method}{Optimisation method. Options: "alternating", "joint",
"stepwise"}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
Estimate a multivariate probit model from multivariate binary
data in a Bayesian generalised linear model framework
}
\details{
For details on the multivariate probit model, see
\code{\link[=mp_model]{mp_model()}}. The \code{multiprobit} function estimates
the model for observations stored in \code{response}
}
\examples{
\dontrun{
if (interactive()) {
  # EXAMPLE1
}
}
}
