% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/environment.R
\name{multiprobit_env_get}
\alias{multiprobit_env_get}
\title{Get access to the internal environment}
\usage{
multiprobit_env_get()
}
\description{
Get access to the internal environment
}
\details{
The environment is defined in aaaaa.R which is loaded first.
}
\keyword{internal}
