% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/environment.R
\name{print.mp_options}
\alias{print.mp_options}
\title{Print multiprobit options}
\usage{
\method{print}{mp_options}(x, legend = TRUE, include_global = TRUE, include_default = TRUE, ...)
}
\arguments{
\item{x}{An \code{\link{mp_options}} object to be printed}

\item{legend}{logical; If \code{TRUE}, include explanatory text, Default: \code{TRUE}}

\item{include_global}{logical; If \code{TRUE}, include global override options}

\item{include_default}{logical; If \code{TRUE}, include default options}

\item{...}{Further parameters, currently ignored}
}
\description{
Print multiprobit options
}
\examples{
if (interactive()) {
  options <- mp_options(verbose = TRUE)

  # Don't print options only set in default:
  print(options, include_default = FALSE)

  # Only include options set in the object:
  print(options, include_default = FALSE, include_global = FALSE)
}

}
