% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimation.R
\name{print.mp_estimate_summary}
\alias{print.mp_estimate_summary}
\alias{print.mp_model_summary}
\title{Print multiprobit objects}
\usage{
\method{print}{mp_estimate_summary}(x, ...)

\method{print}{mp_model_summary}(x, ...)
}
\arguments{
\item{x}{an \code{\link{mp_model}} or \code{\link{mp_estimate_summary}} object to be printed.}

\item{...}{further arguments passed to or from other methods.}
}
\description{
Print multiprobit objects
}
