% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posterior.R
\name{sqrt_expectation}
\alias{sqrt_expectation}
\title{FUNCTION_TITLE}
\usage{
sqrt_expectation(expectation, post_scaling = 1)
}
\arguments{
\item{expectation}{a vector or data frame including variables \code{value} and
\code{std_err}}

\item{post_scaling}{An optional multiplicative scaling to be applied after
the square root operation, Default: 1}
}
\value{
The input \code{expectation} object with the \code{value} and \code{std_err}
elements modified to reflect a square root operation, see Details.
}
\description{
FUNCTION_DESCRIPTION
}
\details{
Matches the mean (the value) and standard deviation (for the
standard error) of a positive random variable to those of a log-Normal
distribution, and computes the expectation and standard deviation of the
square root of that log-Normal distribution.
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\keyword{internal}
