% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posterior.R
\name{update_integration}
\alias{update_integration}
\title{FUNCTION_TITLE}
\usage{
update_integration(integration, weight, quantity, weight_offset = 0)
}
\arguments{
\item{integration}{A \code{list}, or \code{NULL}}

\item{weight}{A scalar weight, either normalised or unnormalised}

\item{quantity}{PARAM_DESCRIPTION}

\item{weight_offset}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESC
\deqn{S_a = \sum (w^a-w_o^a)}
\deqn{S_{a,1} = \sum (w^a-w_o^a) v_m / S_a}
\deqn{S_{a,2} = \sum (w^a-w_o^a) (v_m - S_{a,1})^2 / S_a}
\deqn{S_{a,3} = \sum (w^a-w_o^a) (v_m - S_{a,1})^3 / S_a}
\deqn{S_{a,4} = \sum (w^a-w_o^a) (v_m - S_{a,1})^4 / S_a}
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
  w <- runif(10)
  x <- matrix(rnorm(50), 5, 10)
  int <- NULL
  for (k in seq_along(w)) {
    int <- update_integration(int, w[k], x[,k])
  }

  Alter some weights:
  w2 <- w
  w2[1:2] <- runif(2)
  for (k in which(w2 != w)) {
    int <- update_integration(int, w2[k], x[k], w[k])
  }
}
}
}
