% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{cvec}
\alias{cvec}
\alias{rvec}
\title{Vectorisation utitilies}
\usage{
cvec(A, d = NULL)

rvec(A, d = NULL)
}
\arguments{
\item{A}{A matrix or a vector}

\item{d}{The number of matrix columns}
}
\value{
If \code{A} is a matrix, \code{cvec} returns the columnwise vectorisation of \code{A}.
and \code{rvec} returns the rowwise vectorisation of \code{A}.
If \code{A} is a vector, \code{cvec} and \code{rvec} return a matrix with \code{d} columns,
filled columnwise for \code{cvec} and rowwise for \code{rvec}.
}
\description{
Convert between matrix and vector representations
}
\examples{
if (interactive()) {
  Ac <- cvec(1:6, 2) # As matrix(1:6, 3, 2, byrow = FALSE)
  Ar <- rvec(1:6, 2) # As matrix(1:6, 3, 2, byrow = TRUE)
  cvec(Ac) # As as.vector(Ac)
  rvec(Ar) # As as.vector(t(Ar))
}
}
