% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform.R
\name{wm_model}
\alias{wm_model}
\alias{wm_latent}
\alias{wm_chol}
\alias{wm_matrix}
\alias{wm_latent.wm_model}
\alias{wm_chol.wm_model}
\alias{wm_matrix.wm_model}
\title{Wishart model wrapper}
\usage{
wm_model(type, V = NULL, df = NULL, V_chol = NULL, lower_chol = FALSE)

wm_latent(model, ...)

wm_chol(model, ...)

wm_matrix(model, ...)

\method{wm_latent}{wm_model}(model, W = NULL, W_chol = NULL, lower_chol = NULL, ...)

\method{wm_chol}{wm_model}(model, x, lower_chol = NULL, ...)

\method{wm_matrix}{wm_model}(model, x, ...)
}
\arguments{
\item{type}{DOC}

\item{V}{DOC}

\item{df}{DOC}

\item{V_chol}{DOC}

\item{lower_chol}{DOC}

\item{model}{PARAM_DESCRIPTION}

\item{...}{PARAM_DESCRIPTION}

\item{W}{DOC}

\item{W_chol}{DOC}

\item{x}{DOC}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
