% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform.R
\name{dwishart}
\alias{dwishart}
\title{Wishart Density}
\usage{
dwishart(
  W = NULL,
  x = NULL,
  W_chol = NULL,
  V_chol,
  df,
  lower_chol = FALSE,
  log = FALSE
)
}
\arguments{
\item{W}{A Wishart matrix, Default: NULL}

\item{x}{A latent parameter vector for Wishart model, Default: NULL}

\item{W_chol}{The Cholesky factor of a Wishart matrix, Default: NULL}

\item{V_chol}{The Cholesky factor of the V-parameter for a Wishart model}

\item{df}{The Wishart degrees of freedom parametre}

\item{lower_chol}{\code{TRUE} if lower triangular Cholesky factors are used,
Default: FALSE}

\item{log}{If \code{TRUE}, return the log-density, Default: FALSE}
}
\value{
The density or log-density (if \code{log == TRUE}) for \code{W}, or the \code{W}
matrix constructed from a Cholesky factor \code{W_chol}
}
\description{
Calculate the density for a Wishart model
}
\details{
This function requires the \code{CholWishart::lmvgamma} function from
the \code{CholWishart} package.
}
\note{
This is an internal non-exported function
}
\examples{
\dontrun{
if (interactive()) {
  # EXAMPLE1
}
}
}
\keyword{internal}
