% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform.R
\name{latent_from_nwishart}
\alias{latent_from_nwishart}
\title{Transform Normalised Wishart to latent variables}
\usage{
latent_from_nwishart(LC, LV, df)
}
\arguments{
\item{LC}{The lower Cholesky factor of the Normalised Wishart matrix}

\item{LV}{The lower Cholesky factor of the Wishart $V$ parameter}

\item{df}{The Wishart degrees of freedom}
}
\value{
A list
}
\description{
Transform a Normalised Wishart matrix to latent iid
$N(0,1)$ variables
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\author{
Finn Lindgren
}
