% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/probabilities.R
\name{mpp_gradient_mu}
\alias{mpp_gradient_mu}
\alias{mpp_hessian_mu}
\alias{mpp_gradient_u}
\title{Multivariate Probit Event Probability Derivatives}
\usage{
mpp_gradient_mu(y, mu, ..., seed = 1L, h = 1e-06, symmetric = FALSE)

mpp_hessian_mu(y, mu, ..., seed = 1L, h = 1e-04, diagonal = FALSE)

mpp_gradient_u(
  y,
  mu,
  u,
  V_chol,
  df,
  lower_chol = FALSE,
  seed = 1L,
  h = 1e-06,
  symmetric = FALSE,
  log = FALSE,
  ...
)
}
\arguments{
\item{y}{PARAM_DESCRIPTION}

\item{mu}{PARAM_DESCRIPTION}

\item{...}{Further parameters passed on to
\code{\link{mpp}}}

\item{seed}{The random seed for \code{excursions::gaussint}, Default: NULL}

\item{h}{Step size for finite differences, Default: 1e-06 (for gradients)
or 1e-04 (for hessian)}

\item{symmetric}{For gradients, whether to use symmetric finite differences,
Default: FALSE}

\item{diagonal}{Logical; if \code{TRUE}, only the diagonal of the hessian
is evaluated, Default: FALSE}

\item{u}{A vector of latent variables identifying the Normalised Wishart
matrix, length \eqn{d(d-1)/2}}

\item{V_chol}{The Cholesky factor of the Wishart \eqn{V} parameter}

\item{df}{The Wishart degrees of freedom}

\item{lower_chol}{\code{TRUE} if lower triangular Cholesky factors are used,
Default: \code{FALSE})}

\item{log}{Whether to compute gradient of the log-probability,
Default: \code{FALSE}}
}
\value{
OUTPUT_DESCRIPTION gradient

OUTPUT_DESCRIPTION hessian

OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS gradient

DETAILS hessain

DETAILS
}
\examples{
\dontrun{
if (interactive()) {
  # EXAMPLE1 gradient
}
}
\dontrun{
if (interactive()) {
  # EXAMPLE1 hessian
}
}
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link{mpp}}
}
