% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimation.R
\name{multiprobit}
\alias{multiprobit}
\title{Estimate a multivariate probit model}
\usage{
multiprobit(
  response,
  X = NULL,
  formula = NULL,
  data = NULL,
  df,
  prec_beta = 0,
  gaussint_options = NULL,
  optim_options = NULL,
  max_iter = 10,
  method = c("alternating", "joint", "stepwise")
)
}
\arguments{
\item{response}{A matrix with n-by-d elements, where each row is a multivariate
observation, see Details. A vector is interpreted as a single row matrix.}

\item{X}{An optimally precomputed n-by-J model matrix, where J is the number
regression coeficcients for each of the d dimensions.}

\item{formula}{A formula interpretable by \code{model.matrix}.}

\item{data}{A \code{data.frame} containing the veriables needed by the
formula.}

\item{df}{Degrees of freedom for the normalised Wishart prior for the
correlation matrix. See Details.}

\item{prec_beta}{Prior precision for the regression coefficients}

\item{gaussint_options}{Optional list of options for
  \code{excursions::gaussint}.
  Specific relevant option:
  \code{num_threads} The maximum number of allowed threads for parallel
computing by \code{excursions::gaussint}, Default: 0, meaning no limit.}

\item{optim_options}{Optional list of control options for \code{optim()}}

\item{max_iter}{The maximum number of steps for
\code{method == "alternating"}}

\item{method}{Optimisation method. Options: "alternating", "joint",
"stepwise"}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
Estimate a multivariate probit model from multivariate binary
  data in a Bayesian generalised linear model framework
}
\details{
Estimates a multivariate probit model for the response variable,
here denoted \eqn{Y}. The model is built from a linear predictor
\deqn{M = X B}
where \eqn{X} is a n-by-J matrix of \eqn{J} predictors, and \eqn{B} is
a J-by-d matrix of regression coefficients.
Each row of \eqn{M} is the linear
predictor for one multivariate observation. The response variables \eqn{Y}
are linked to \eqn{M} by first defining latent Gaussian variables
\deqn{Z=M+E} where each row of \eqn{E} is a multivariate Normal vector,
\eqn{E ~ N(0,\Sigma)}. Then, \deqn{Y_{i,k}=I(Z_{i,k} > 0).}
Conditionally on \eqn{B}, each row of \eqn{Y} has a multinomial distribution
on the set of all \eqn{0/1} combinations, with each probability equal to a
hyperquadrant probability of a the multivariate Normal distribution
\eqn{N(\mu,\Sigma)}, where \eqn{\mu} is the corresponding row of \eqn{M}.
Only the inequality \eqn{Y_{i,k} > 0} for the response variables is used,
so alternative data representations such as \eqn{-1/+1} will also work as
expected.

The degrees of freedom for the normalised Wishart prior are linked to the
LKJ prior by \code{df = ...}, which makes the two models equivalent.
}
\examples{
\dontrun{
if (interactive()) {
  # EXAMPLE1
}
}
}
