% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimation.R
\name{multiprobit}
\alias{multiprobit}
\title{Estimate a multivariate probit model}
\usage{
multiprobit(model = NULL, ..., options = NULL)
}
\arguments{
\item{model}{An \code{\link{mp_model}} model object.}

\item{...}{Parameters passed on to \code{\link[=mp_model]{mp_model()}}}

\item{options}{An \code{\link{mp_options}} object or a list that can be coerced into
an \code{mp_options} object. Options set here will override the global options.}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
Estimate a multivariate probit model from multivariate binary
data in a Bayesian generalised linear model framework
}
\details{
For details on the multivariate probit model, see
\code{\link[=mp_model]{mp_model()}}. The \code{multiprobit} function estimates
the model for observations stored in \code{response}
}
\examples{
\dontrun{
if (interactive()) {
    N <- 6
    d <- 2
    J <- 1

    set.seed(1L)
    X <- cbind(1, matrix(rnorm(N * (J - 1)), N, J - 1))
    B <- matrix(0.5, J, d)
    Y <- matrix(rnorm(N * d, mean = as.vector(X \%*\% B)) > 0, N, d)
    df <- d + 1
    prec_beta <- 0.1

    model <- mp_model(response = Y, X = X,
                      df = df, prec_beta = prec_beta)
    opt <- multiprobit(model = model,
                       options =
                       mp_options(
                       gaussint = list(max.threads = 1),
                       strategy = "stepwise"))
}
}
}
