% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{tri_solve}
\alias{tri_solve}
\alias{qchisq_pnorm}
\alias{qnorm_pchisq}
\alias{qbeta_pchisq}
\alias{qchisq_pbeta}
\alias{sparse_identity}
\title{Internal multiprobit utilities}
\usage{
tri_solve(A, b, lower_tri = FALSE)

qchisq_pnorm(x, df)

qnorm_pchisq(x, df)

qbeta_pchisq(x, df, shape1, shape2, ncp)

qchisq_pbeta(x, shape1, shape2, df, ncp)

sparse_identity(d)
}
\description{
Internal multiprobit utilities
}
\details{
\code{tri_solve} solves triangular systems with back/forwardsolve

\code{qchisq_pnorm} evaluates \code{qchisq(pnorm(...))} with
attempt at numerical stability.
}
\keyword{internal}
