% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimation.R
\name{print.mp_estimate_summary}
\alias{print.mp_estimate_summary}
\alias{print.mp_model_summary}
\title{Print multiprobit objects}
\usage{
\method{print}{mp_estimate_summary}(object, ...)

\method{print}{mp_model_summary}(object, ...)
}
\description{
Print multiprobit objects
}
