% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimation.R
\name{summary.mp_estimate}
\alias{summary.mp_estimate}
\title{FUNCTION_TITLE}
\usage{
\method{summary}{mp_estimate}(object, ...)
}
\arguments{
\item{object}{PARAM_DESCRIPTION}

\item{...}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
