% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimation.R
\name{summary.mp_estimate}
\alias{summary.mp_estimate}
\title{Multiprobit Estimate Summary Information}
\usage{
\method{summary}{mp_estimate}(object, ...)
}
\arguments{
\item{object}{An \code{\link{mp_estimate}} object}

\item{...}{Additional parameters, currently unused.}
}
\value{
An \code{mp_estimate_summary} object
}
\description{
Organise information about multiprobit parameter estimates
}
\examples{
\dontrun{
if (interactive()) {
  # EXAMPLE1
}
}

}
